Tog bort data filen helt. Flyttade ut cartitem Interfaces som låg där till cartcontext
