--- Client/src/data/index.ts
@@ -1,8 +1,0 @@
-export interface CartItem {
-  id: string;
-  title: string;
-  price: number;
-  quantity: number;
-  imageID: string;
-}
-
